Avoid using ErrorReturnObj
In GAP 4.11, most uses of ErrorReturnObj were removed, and in GAP 4.12
all of them. Background is that we found that it is extremely hard for
users to use them correctly, but incorrect use quickly leads to crashes,
or worse (nonsense results of computations). Based on some informal
polling, we found nobody who actually uses this feature other than in
extremely rare fringe cases, and nobody actually expressed any regret
about removing this feature.

Note that ErrorReturnVoid for now still is possible, there are a few
cases where it's useful, but we also suggest to minimize usage of that.